[Refactor]: Split RTK endpoints/services
