+++ cardstack/src/components/TransactionConfirmationSheet/displays/components/sections/PrepaidCardTransactionSection.tsx
@@ -5,7 +5,7 @@ import TransactionListItem from '../TransactionListItem';
 import { Container, Text } from '@cardstack/components';
 import { useNativeCurrencyAndConversionRates } from '@rainbow-me/redux/hooks';
 import { convertSpendForBalanceDisplay } from '@cardstack/utils';
-import { useGetSafesDataQuery } from '@cardstack/services/safes';
+import { useGetSafesDataQuery } from '@cardstack/services';
 import { useAccountSettings } from '@rainbow-me/hooks';
 
 export const PrepaidCardTransactionSection = ({
