+++ cardstack/src/models/safes-providers.ts
@@ -3,8 +3,6 @@ import { captureException } from '@sentry/minimal';
 import { SignedProviderParams } from './hd-provider';
 import Web3Instance from './web3-instance';
 import logger from 'logger';
-import { MainRoutes } from '@cardstack/navigation';
-import { Navigation } from '@rainbow-me/navigation';
 
 export const getSafesInstance = async (
   signedProviderParams?: SignedProviderParams
@@ -15,7 +13,6 @@ export const getSafesInstance = async (
 
     return safes;
   } catch (e) {
-    Navigation.handleAction(MainRoutes.ERROR_FALLBACK_SCREEN, {}, true);
     captureException(e);
     logger.sentry('Unable to get safeInstance', e);
   }
